Allow worker.js to work with pyodide.mjs
If you build this as an es module with webpack, it breaks because importScripts doesn't work in ES modules, and the worker module is one. This change gives the option to pass a link to pyodide.mjs (the es module version of pyodide) which means the whole of jupyterlite can happily build as es modules.

I added a `webpackIgnore` comment because it is easier to leave pyodide as a completely external dependency and not have to mess around with adding it into the webpack. It would be possible to use NPM embedded pyodide, but that would up the build size of jupyterlite.