Allow to install extra packages from extra indexes from jupyter-lite.json<F48>
This is in progress, but the idea is to be able to point to dev wheels
on rtd or similar without having to make a custom pyidide distribution.

For example:

```
// jupyter-lite.json
{
  "jupyter-config-data": {
    "litePluginSettings": {
       "extraPackagesAndIndexes": [
          {
            "indexes": [
              "https://cors.carreau.workers.dev/scientific-python-nightly-wheels/simple",
              "https://pypi.org/simple"
            ],
            "package": "xarray"
          }
        ]
      }
    }
  }
```

Package could also be a simple url to a whl (I need to check how it
works if it's a local URL), and indexes can be ommited.

WIP as I'm stuggling to have the typescript build process work.