WIP, publish error tracebacks on iopub
This ensures that when an Output ipywidget is used as a context manager,
errors raised inside the context are rendered in the Output widget.

```
out = Output()
out

with out:
    raise RuntimeError("oh no! an error occurred")
```

However, it does not suppress the error in the original cell, which is
to say the error is rendered both in the original cell and in the Output
widget.  This behavior is in contrast to Jupyter Lab and Notebook, which
do suppress the error in the original cell.  Fixing this is a TODO.

This patch also handles the case where an ipywidgets handler, without
the `@out.capture()` decorator, raises an exception.  Such exceptions
are now seen in the log console.

TODOs:
- [ ] Suppress error in the original cell